% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-caviar_end.R
\docType{data}
\name{caviar_end}
\alias{caviar_end}
\title{Caviar end data}
\format{
A dataframe with 72 rows and 3 variables:
\describe{
\item{from}{An individual under surveillance}
\item{to}{An individual under surveillance}
\item{weight}{The number of intercepted communications between the individuals}
}
}
\source{
\href{https://link.springer.com/book/10.1007/978-0-387-09526-4}{Carlo Morselli}
}
\usage{
caviar_end
}
\description{
Network edgelist data as at the end of the Operation Caviar investigation into drug trafficking in Canada
}
\examples{
caviar_end
}
\keyword{datasets}
