% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-chinook_customers.R
\docType{data}
\name{chinook_customers}
\alias{chinook_customers}
\title{Chinook customer data}
\format{
A dataframe with 59 rows and 4 variables:
\describe{
\item{CustomerId}{Customer ID number}
\item{FirstName}{Customer First Name}
\item{LastName}{Customer Last Name}
\item{SupportRepId}{ID of Sales Rep assigned to customer}
}
}
\source{
\href{https://github.com/lerocha/chinook-database}{Chinook Database on Github}
}
\usage{
chinook_customers
}
\description{
Extract of data on customers of a music sales company
}
\examples{
chinook_customers
}
\keyword{datasets}
