% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-chinook_employees.R
\docType{data}
\name{chinook_employees}
\alias{chinook_employees}
\title{Chinook employee data}
\format{
A dataframe with 8 rows and 4 variables:
\describe{
\item{EmployeeId}{Employee ID number}
\item{FirstName}{Employee First Name}
\item{LastName}{Employee Last Name}
\item{ReportsTo}{ID of Employee who they report to}
}
}
\source{
\href{https://github.com/lerocha/chinook-database}{Chinook Database on Github}
}
\usage{
chinook_employees
}
\description{
Extract of data on employees of a music sales company
}
\examples{
chinook_employees
}
\keyword{datasets}
