% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-chinook_invoices.R
\docType{data}
\name{chinook_invoices}
\alias{chinook_invoices}
\title{Chinook invoice data}
\format{
A dataframe with 412 rows and 2 variables:
\describe{
\item{InvoiceId}{Invoice ID number}
\item{CustomerId}{CustomerID number}
}
}
\source{
\href{https://github.com/lerocha/chinook-database}{Chinook Database on Github}
}
\usage{
chinook_invoices
}
\description{
Extract of data on customer invoices from a music sales company
}
\examples{
chinook_invoices
}
\keyword{datasets}
