% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-chinook_items.R
\docType{data}
\name{chinook_items}
\alias{chinook_items}
\title{Chinook item sales data}
\format{
A dataframe with 2240 rows and 2 variables:
\describe{
\item{InvoiceId}{ID number of invoice containing the item}
\item{TrackId}{ID number of the item}
}
}
\source{
\href{https://github.com/lerocha/chinook-database}{Chinook Database on Github}
}
\usage{
chinook_items
}
\description{
Extract of data on items sold by a music sales company
}
\examples{
chinook_items
}
\keyword{datasets}
