% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dolphins.R
\docType{data}
\name{dolphins}
\alias{dolphins}
\title{Bottlenose dolphin social network}
\format{
A dataframe with 159 rows and 2 variables:
\describe{
\item{from}{Dolphin ID}
\item{to}{Dolphin ID}
}
}
\source{
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC1809954/pdf/14667378.pdf}{Lusseau et al, 2003}
}
\usage{
dolphins
}
\description{
Edgelist of network of frequent interaction between bottlenose dolphins in Doubtful Sound, New Zealand
}
\examples{
dolphins
}
\keyword{datasets}
