% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-email_edgelist.R
\docType{data}
\name{email_edgelist}
\alias{email_edgelist}
\title{Email edgelist}
\format{
A dataframe with 24929 rows and 2 variables:
\describe{
\item{from}{ID of sender}
\item{to}{ID of receiver}
}
}
\source{
\href{http://www.cs.cmu.edu/~jure/pubs/powergrowth-tkdd.pdf}{Leskovec et al, 2007}
}
\usage{
email_edgelist
}
\description{
Edgelist of network of email communications at a large European research institution
}
\examples{
email_edgelist
}
\keyword{datasets}
