% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-email_vertices.R
\docType{data}
\name{email_vertices}
\alias{email_vertices}
\title{Email vertices}
\format{
A dataframe with 1005 rows and 2 variables:
\describe{
\item{id}{Vertex ID of individual}
\item{dept}{Department of individual}
}
}
\source{
\href{http://www.cs.cmu.edu/~jure/pubs/powergrowth-tkdd.pdf}{Leskovec et al, 2007}
}
\usage{
email_vertices
}
\description{
Vertex data of network of email communications at a large European research institution
}
\examples{
email_vertices
}
\keyword{datasets}
