% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-eu_referendum.R
\docType{data}
\name{eu_referendum}
\alias{eu_referendum}
\title{EU referendum data}
\format{
A dataframe with 382 rows and 4 variables:
\describe{
\item{Region}{UK Region}
\item{Area_Code}{UK Area Code}
\item{Remain}{Number of votes to remain in the EU}
\item{Leave}{Number of votes to leave the EU}
}
}
\source{
\href{https://www.electoralcommission.org.uk/who-we-are-and-what-we-do/elections-and-referendums/past-elections-and-referendums/eu-referendum/results-and-turnout-eu-referendum}{UK Electoral Commission}
}
\usage{
eu_referendum
}
\description{
Data on voting in the UK EU membership referendum in 2016
}
\examples{
eu_referendum
}
\keyword{datasets}
