% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-friends_tv_edgelist.R
\docType{data}
\name{friends_tv_edgelist}
\alias{friends_tv_edgelist}
\title{Friends TV edgelist}
\format{
A dataframe with 2976 rows and 3 variables:
\describe{
\item{from}{Friends character}
\item{to}{Friends character}
\item{weight}{Number of scenes with both characters}
}
}
\source{
\href{https://github.com/keithmcnulty/friends_analysis}{Keith McNulty}
}
\usage{
friends_tv_edgelist
}
\description{
Edgelist of network of characters of US TV Show Friends based on appearing in the same scene
}
\examples{
friends_tv_edgelist
}
\keyword{datasets}
