% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-g14_edgelist.R
\docType{data}
\name{g14_edgelist}
\alias{g14_edgelist}
\title{G14 edgelist}
\format{
A dataframe with 18 rows and 3 variables:
\describe{
\item{from}{Vertex ID}
\item{to}{Vertex ID}
\item{weight}{Edge weight}
}
}
\usage{
g14_edgelist
}
\description{
Edgelist of small network of 14 vertices
}
\examples{
g14_edgelist
}
\keyword{datasets}
