% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-karate.R
\docType{data}
\name{karate}
\alias{karate}
\title{Zachary's Karate Club edgelist}
\format{
A dataframe with 78 rows and 2 variables:
\describe{
\item{from}{Member ID}
\item{to}{Member ID}
}
}
\source{
\href{https://www.jstor.org/stable/3629752?origin=JSTOR-pdf}{Zachary, 1977}
}
\usage{
karate
}
\description{
Edgelist of network of social interactions between members of a karate club
}
\examples{
karate
}
\keyword{datasets}
