% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-koenigsberg.R
\docType{data}
\name{koenigsberg}
\alias{koenigsberg}
\title{Bridges of Koenigsberg edgelist}
\format{
A dataframe with 7 rows and 2 variables:
\describe{
\item{from}{Place name}
\item{to}{Place name}
}
}
\source{
\href{https://archive.org/details/commentariiacade08impe/page/128/mode/2up}{Euler, 1736}
}
\usage{
koenigsberg
}
\description{
Edgelist of network of places connected by bridges in the city of Koenigsberg
}
\examples{
koenigsberg
}
\keyword{datasets}
