% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lesmis.R
\docType{data}
\name{lesmis}
\alias{lesmis}
\title{Les Miserables character network}
\format{
A dataframe with 254 rows and 3 variables:
\describe{
\item{from}{Character name}
\item{to}{Character name}
\item{weight}{Number of chapters both characters appear in}
}
}
\source{
\href{https://dl.acm.org/doi/book/10.1145/164984}{Knuth, 1993}
}
\usage{
lesmis
}
\description{
Edgelist of network of characters in Victor Hugo's Les Miserables based on appearance in the same chapter
}
\examples{
lesmis
}
\keyword{datasets}
