% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-londontube_edgelist.R
\docType{data}
\name{londontube_edgelist}
\alias{londontube_edgelist}
\title{London Tube network edgelist}
\format{
A dataframe with 406 rows and 4 variables:
\describe{
\item{from}{Station ID}
\item{to}{Station ID}
\item{line}{Name of line connecting stations}
\item{linecolor}{Official color of line connecting stations}
}
}
\usage{
londontube_edgelist
}
\description{
Edgelist of network of London Tube/Underground stations
}
\examples{
londontube_edgelist
}
\keyword{datasets}
