% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-londontube_vertices.R
\docType{data}
\name{londontube_vertices}
\alias{londontube_vertices}
\title{London Tube network vertices}
\format{
A dataframe with 302 rows and 4 variables:
\describe{
\item{id}{Station ID}
\item{name}{Station name}
\item{latitude}{Station latitude}
\item{longitude}{Station longitude}
}
}
\usage{
londontube_vertices
}
\description{
Vertices of network of London Tube/Underground stations
}
\examples{
londontube_vertices
}
\keyword{datasets}
