% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-madmen_edges.R
\docType{data}
\name{madmen_edges}
\alias{madmen_edges}
\title{Mad Men network edgelist}
\format{
A dataframe with 39 rows and 3 variables:
\describe{
\item{Name1}{Character name}
\item{Name2}{Character name}
\item{Married}{Whether the relationship was part of a marriage}
}
}
\usage{
madmen_edges
}
\description{
Edgelist of network of romantic relationships between characters of the TV show Mad Men
}
\examples{
madmen_edges
}
\keyword{datasets}
