% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-madmen_vertices.R
\docType{data}
\name{madmen_vertices}
\alias{madmen_vertices}
\title{Mad Men network vertices}
\format{
A dataframe with 45 rows and 3 variables:
\describe{
\item{label}{Character name}
\item{Gender}{Character gender}
\item{Main}{Whether the character is a main character}
}
}
\usage{
madmen_vertices
}
\description{
Vertices of network of romantic relationships between characters of the TV show Mad Men
}
\examples{
madmen_vertices
}
\keyword{datasets}
