% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-netscience.R
\docType{data}
\name{netscience}
\alias{netscience}
\title{Network Science collaboration network}
\format{
A dataframe with 2742 rows and 3 variables:
\describe{
\item{from}{Scientist name}
\item{to}{Scientist name}
\item{weight}{Measure of strength of collaboration}
}
}
\source{
\href{https://journals.aps.org/pre/abstract/10.1103/PhysRevE.74.036104}{Newman, 2006}
}
\usage{
netscience
}
\description{
Edgelist of network of academic collaboration between network scientists
}
\examples{
netscience
}
\keyword{datasets}
