% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ontariopol_edgelist.R
\docType{data}
\name{ontariopol_edgelist}
\alias{ontariopol_edgelist}
\title{Ontario politician Twitter interaction network edglist}
\format{
A dataframe with 6095 rows and 3 variables:
\describe{
\item{from}{Politician ID}
\item{to}{Politician ID}
\item{weight}{Number of Twitter interactions}
}
}
\source{
\href{https://cbelanger.netlify.app/post/ontario-meet-your-mpps-most-prolific-posters/}{Christopher Belanger}
}
\usage{
ontariopol_edgelist
}
\description{
Edgelist of Twitter interaction network of Ontario province politicians
}
\examples{
ontariopol_edgelist
}
\keyword{datasets}
