% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ontariopol_vertices.R
\docType{data}
\name{ontariopol_vertices}
\alias{ontariopol_vertices}
\title{Ontario politician Twitter interaction network vertices}
\format{
A dataframe with 108 rows and 4 variables:
\describe{
\item{id}{Politician ID}
\item{screen_name}{Politician Twitter screen name}
\item{name}{Politician name}
\item{party}{Party affiliation}
}
}
\source{
\href{https://cbelanger.netlify.app/post/ontario-meet-your-mpps-most-prolific-posters/}{Christopher Belanger}
}
\usage{
ontariopol_vertices
}
\description{
Vertices of Twitter interaction network of Ontario province politicians
}
\examples{
ontariopol_vertices
}
\keyword{datasets}
