% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-park_reviews.R
\docType{data}
\name{park_reviews}
\alias{park_reviews}
\title{Yelp park reviews}
\format{
A dataframe with 231 rows and 4 variables:
\describe{
\item{park_id}{Park ID}
\item{user_id}{User ID}
\item{park_name}{Park name}
\item{stars}{Number of stars awarded by user}
}
}
\usage{
park_reviews
}
\description{
Data on Yelp reviews of dog parks in Phoenix, AZ
}
\examples{
park_reviews
}
\keyword{datasets}
