% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-s50_edges.R
\docType{data}
\name{s50_edges}
\alias{s50_edges}
\title{Teenage Friends and Lifestyle Study network edgelist}
\format{
A dataframe with 122 rows and 2 variables:
\describe{
\item{from}{Person ID}
\item{to}{Person ID}
}
}
\source{
\href{https://www.tandfonline.com/doi/abs/10.1080/dep.7.1.21.37}{Pearson & Michell, 2009}
}
\usage{
s50_edges
}
\description{
Edgelist of friend network of teenage girls in Scotland
}
\examples{
s50_edges
}
\keyword{datasets}
