% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-s50_vertices.R
\docType{data}
\name{s50_vertices}
\alias{s50_vertices}
\title{Teenage Friends and Lifestyle Study network vertices}
\format{
A dataframe with 50 rows and 5 variables:
\describe{
\item{id}{Person ID}
\item{smoke}{Frequency of smoking from 1 (Never) to 3 (Regularly)}
\item{alcohol}{Frequency of drinking alcohol from 1 (Never) to 5 (More than once a week)}
\item{drugs}{Frequency of cannabis use from 1 (Never) to 4 (Regularly)}
\item{sport}{Frequency of sporting activity from 1 (Not regularly) to 2 (Regularly)}
}
}
\source{
\href{https://www.tandfonline.com/doi/abs/10.1080/dep.7.1.21.37}{Pearson & Michell, 2009}
}
\usage{
s50_vertices
}
\description{
Vertices of friend network of teenage girls in Scotland
}
\examples{
s50_vertices
}
\keyword{datasets}
