% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-schoolfriends_edgelist.R
\docType{data}
\name{schoolfriends_edgelist}
\alias{schoolfriends_edgelist}
\title{Schoolfriends network edgelist}
\format{
A dataframe with 2105 rows and 3 variables:
\describe{
\item{from}{Person ID}
\item{to}{Person ID}
\item{type}{Whether the friendship is a known Facebook connection or if it was reported by \code{from} person}
}
}
\source{
\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0136497}{Mastandrea et al, 2015}
}
\usage{
schoolfriends_edgelist
}
\description{
Edgelist of network of schoolfriends in a French high school
}
\examples{
schoolfriends_edgelist
}
\keyword{datasets}
