% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-schoolfriends_vertices.R
\docType{data}
\name{schoolfriends_vertices}
\alias{schoolfriends_vertices}
\title{Schoolfriends network vertices}
\format{
A dataframe with 329 rows and 3 variables:
\describe{
\item{id}{Person ID}
\item{class}{School class of person}
\item{gender}{Gender of person}
}
}
\source{
\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0136497}{Mastandrea et al, 2015}
}
\usage{
schoolfriends_vertices
}
\description{
Vertices of network of schoolfriends in a French high school
}
\examples{
schoolfriends_vertices
}
\keyword{datasets}
