% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wikivote.R
\docType{data}
\name{wikivote}
\alias{wikivote}
\title{Wikipedia administrator voting network}
\format{
A dataframe with 103688 rows and 2 variables:
\describe{
\item{from}{ID of voter}
\item{to}{ID of vote recipient}
}
}
\usage{
wikivote
}
\description{
Edgelist of network of votes for Wikipedia administrators
}
\examples{
wikivote
}
\keyword{datasets}
