% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-workfrance_edgelist.R
\docType{data}
\name{workfrance_edgelist}
\alias{workfrance_edgelist}
\title{Workplace network edgelist}
\format{
A dataframe with 932 rows and 3 variables:
\describe{
\item{from}{Person ID}
\item{to}{Person ID}
\item{mins}{Number of minutes spent co-located}
}
}
\source{
\href{https://www.cambridge.org/core/journals/network-science/article/abs/data-on-facetoface-contacts-in-an-office-building-suggest-a-lowcost-vaccination-strategy-based-on-community-linkers/18AB49AB4F2AEA33CE7501F06ADBC8E8}{Génois et al, 2015}
}
\usage{
workfrance_edgelist
}
\description{
Edgelist of network of interactions between people in a French office building based on location sensor technology
}
\examples{
workfrance_edgelist
}
\keyword{datasets}
