% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-workfrance_vertices.R
\docType{data}
\name{workfrance_vertices}
\alias{workfrance_vertices}
\title{Workplace network vertices}
\format{
A dataframe with 211 rows and 2 variables:
\describe{
\item{id}{Person ID}
\item{dept}{Department of person}
}
}
\source{
\href{https://www.cambridge.org/core/journals/network-science/article/abs/data-on-facetoface-contacts-in-an-office-building-suggest-a-lowcost-vaccination-strategy-based-on-community-linkers/18AB49AB4F2AEA33CE7501F06ADBC8E8}{Génois et al, 2015}
}
\usage{
workfrance_vertices
}
\description{
Vertices of network of interactions between people in a French office building based on location sensor technology
}
\examples{
workfrance_vertices
}
\keyword{datasets}
